change datatype

+++ src/MORRF/morrf/KDTree2D.h
@@ -11,7 +11,7 @@ class RRTNode;
 
 class POS2D {
 public:
-    typedef int value_type;
+    typedef double value_type;
 
     POS2D() {
         d[0] = 0;
